refactor(compact): extract AutoCompactToggle into its own component
Pulls the Switch + Tooltip + hooks (admin check, useSession, useToggleSessionAutoCompact)
out of session-chat-column.tsx into a small self-contained file alongside it,
matching the codebase pattern of one-component-per-file in components/layout/.

session-chat-column.tsx now just imports the toggle and drops it into the
existing footer row — the PR diff for that file shrinks from ~76 lines of
mixed layout reflow + new logic to a ~5 line addition.
